refactoring: utils
