+++ src/chmock/destinations.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from '../utils.js';
+import {getRandomInteger} from '../utils/common.js';
 
 const destinationNameList = [
   'Chamonix',
